Added java-config
This is intended to be used by R's R CMD javareconf (and anything
else that needs to dynamically reconfigure java).

It depends only on JAVA_HOME (and on barely functional Unix shells
CONDA_PREFIX too).

If JAVA_HOME is not set it will run the first 'java' found on PATH.